requirements.txt
pandas
numpy
Flask
Flask-Cors
ipykernel
seaborn
scikit-learn==1.2.2
database_connect
pymongo
imblearn
xgboost
python-dotenv
python-box
PyYAML
ensure
joblib
mlflow
-e .